Made Basic Notepad
Made a basic notepad that can have information added to it as the player interacts with different objects. The notepad is opened using the letter n and can be closed using n. It can be opened in the interrogation room allowing for the player to read any information they've found.