Highlighting in the notepad and restart button
The words that appear in the blanked spaces are now bolded and in red to give the player a better indication of what was shown/added to the notepad. A restart button was also added to the interrogation scene so that the player can restart to the crime scene if needed.